+++ server/src/lib/aws/s3.ts
@@ -1,11 +1,28 @@
 import { DeleteObjectCommand, PutObjectCommand, S3Client, S3ClientResolvedConfig, ServiceInputTypes, ServiceOutputTypes } from '@aws-sdk/client-s3';
 import type { Client } from '@smithy/types';
-import { getS3Client } from './iam';
 import { getSignedUrl } from '@aws-sdk/s3-request-presigner';
+import { getCredentials } from './iam';
 
+// env
 const BUCKET_NAME = 'acmwebsite-dev-588738592350-us-west-2';
+const region = 'us-west-2';
 
-const uploadFile = async (file: File): Promise<boolean> => {
+let s3client: S3Client | null = null;
+
+const getS3Client = async (): Promise<S3Client | null> => {
+    const credentials = await getCredentials();
+    if(credentials === null) {
+        return null;
+    } 
+    s3client = new S3Client({region, credentials: {
+        accessKeyId: <string> credentials.AccessKeyId,
+        secretAccessKey: <string> credentials.SecretAccessKey,
+        sessionToken: <string> credentials.SessionToken,    
+    }});
+    return s3client;
+};
+
+const uploadFile = async (file: File, key: string): Promise<boolean> => {
     const s3client: S3Client | null = await getS3Client();
     if(s3client === null) {
         return false;
@@ -15,7 +32,7 @@ const uploadFile = async (file: File): Promise<boolean> => {
             const uploadObjectCommand = new PutObjectCommand(
                 {
                     Bucket: BUCKET_NAME,
-                    Key: file.name,
+                    Key: key,
                     Body: (new Buffer(await file.arrayBuffer())),
                 });
             await (<S3Client>s3client).send(uploadObjectCommand);
